[BugFix] Wait for compute before offloading KV to CPU
This commit fixes the OffloadingConnector to wait on the default (compute) stream,
before offloading KV data from the GPU to the CPU.
Additionally, we move the offloading to start at the beginning of the next engine step.
This is to avoid contention with sample_tokens which also involves gpu->cpu copies.
Lastly, we remove the use of stream priorities as they don't effect DMA-based copies.

Signed-off-by: Or Ozeri <oro@il.ibm.com>